First commit after some hacking to make a proof of concept HAPI FHIR based XML to JSON convertor
This kinda works and actually is quite performant -- IVY and initializing takes up 10-15 seconds, but the actual conversion is done in a blink of an eye. So we need to batch all conversions if we want to continue with this.

This is not a working solution yet. It just uses some hardcoded paths. A main issue is also that the convertor trips up on Touchstone placeholders, so we have to find some way around this.